test(dataset): move Dataset-dependent build_manifest test out of transaction.rs
test_build_manifest_partial_last_updated_rewrite_columns_stable_row_ids is
the only test in transaction.rs that needs Dataset, WriteParams and Session.
The other 69 build their inputs by hand. Moving it to dataset_transactions.rs,
where the dataset-level transaction tests already live, lets transaction.rs
move to lance-table unchanged.
